Działający algorytm LSI
- dane są wstepnie przetwarzane (usunięcie stop words, lematyzacja)
- tworzony jest wektor ilości wystapień danego termu w danym dokumencie
  (wektor dla termów uporządkowanych alfabetycznie)
- wektory skłądane są w macierz, ta nastepnie poddawana jest SVD
- tworzona jest na podstawie macierzy SVD macierz reprezentująca słowa
- dla par słow z pliku testowego obliczene jest kosinusowe podobieństwo
  według algorytmu LSI
- różnice podobieństwa dla par wyrazów zapisywane są do listy
- na podstawie wartości listy dokonywana jest ewaluacja

Do zrobienia :
- rozdzielenie obliczen miedzy aktorów według komentarzy w main
- sprawdzenie czy dostajemy te same wartości co obecnie

+++ src/main/java/postprocessing/QualityMeasureEnum.java
@@ -3,7 +3,7 @@ package postprocessing;
 public enum QualityMeasureEnum {
     GREAT, GOOD, BAD;
 
-    public static QualityMeasureEnum getQualityMeasure(Float value) {
+    public static QualityMeasureEnum getQualityMeasure(Double value) {
         if (value < 1.0)
             return GREAT;
         else if (value < 3.0)
